backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.2.4
cx_Oracle==8.3.0
google-generativeai==0.8.4
python-magic-bin==0.4.14
# PDF 파일 지원
PyPDF2==3.0.1
pdf2image==1.17.0
pdfminer.six>=20240706
# 텍스트 요약
transformers==4.50.0
torch==2.6.0
# 언어 감지 및 번역
langdetect==1.0.9
# googletrans==4.0.2
deep-translator==1.11.4
# 표 데이터 인식
camelot-py==1.0.0
tabula-py==2.10.0
JPype1==1.4.1
openpyxl==3.1.5
# 머신러닝 및 NLP
scikit-learn==1.6.1
spacy==3.8.4
# 캐싱 및 최적화
joblib==1.3.2
lru-dict==1.2.0